requirements.txt
streamlit
openpyxl
sqlite3
python-dotenv
pdfplumber 
pillow 
pytesseract
groq-client
langchain
pydantic
langchain_core
langchain_groq
langchain
unstructured-inference
langchain_community
langchain_upstage
langchain-pymupdf4llm
pdf2image
